[GLUTEN-6251][CH] Disable GlutenSortShuffleSuite in clickhouse backend
