+++ gluten-ut/common/src/test/scala/org/apache/spark/sql/GlutenSQLTestsBaseTrait.scala
@@ -20,35 +20,12 @@ import org.apache.gluten.GlutenConfig
 import org.apache.gluten.utils.{BackendTestUtils, SystemParameters}
 
 import org.apache.spark.SparkConf
-import org.apache.spark.sql.GlutenTestConstants.GLUTEN_TEST
 import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.execution.adaptive.{AdaptiveSparkPlanExec, ShuffleQueryStageExec}
 import org.apache.spark.sql.test.SharedSparkSession
 
-import org.scalactic.source.Position
-import org.scalatest.Tag
-
 /** Basic trait for Gluten SQL test cases. */
 trait GlutenSQLTestsBaseTrait extends SharedSparkSession with GlutenTestsBaseTrait {
-
-  protected def testGluten(testName: String, testTag: Tag*)(testFun: => Any)(implicit
-      pos: Position): Unit = {
-    test(GLUTEN_TEST + testName, testTag: _*)(testFun)
-  }
-
-  protected def ignoreGluten(testName: String, testTag: Tag*)(testFun: => Any)(implicit
-      pos: Position): Unit = {
-    super.ignore(GLUTEN_TEST + testName, testTag: _*)(testFun)
-  }
-
-  override protected def test(testName: String, testTags: Tag*)(testFun: => Any)(implicit
-      pos: Position): Unit = {
-    if (shouldRun(testName)) {
-      super.test(testName, testTags: _*)(testFun)
-    } else {
-      super.ignore(testName, testTags: _*)(testFun)
-    }
-  }
 
   override def sparkConf: SparkConf = {
     GlutenSQLTestsBaseTrait.nativeSparkConf(super.sparkConf, warehouse)
